new command registration system proposal
