+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/tecton-tech/Discord-Bot#readme",
   "dependencies": {
+    "annotations": "github:voidedWarranties/annotations",
     "axios": "^0.18.1",
     "body-parser": "^1.19.0",
     "cryptiles": ">=4.1.2",
